adjusted pve rounds & opponent title
