+++ app/models/colyseus-models/player.ts
@@ -37,6 +37,7 @@ export default class Player extends Schema implements IPlayer {
   @type("uint8") interest: number = 0
   @type("string") opponentName: string = ""
   @type("string") opponentAvatar: string = ""
+  @type("string") opponentTitle: string = ""
   @type("uint8") boardSize: number = 0
   @type({ collection: "string" }) items = new CollectionSchema<Item>()
   @type("uint8") rank: number
